dom/credentialmanagement/identity/tests/mochitest/test_get_without_providers.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>No Providers Specified</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="head.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>
    SimpleTest.waitForExplicitFinish();
    setupTest("get_without_providers").then(
      function () {
        return navigator.credentials.get({
          identity: {
          }
        });
      }
    ).then((cred) => {
      ok(false, "incorrectly got a credential");
    }).catch((err) => {
      ok(true, "correctly got an error");
    }).finally(() => {
      SimpleTest.finish();
    })
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">This test verifies that we do not get a credential when we give no providers field in the JSON. This is mostly to make sure we don't have any nullptr derefs.</div>
<pre id="test"></pre>
</body>
</html>
